Cleaning up CcdAssembler to use common interface.
High-level tasks should be agnostic as to whether a Ccd is represented by a
single or multiple exposures.  To this end, reworked CcdAssembler to present
a common interface that covers both options; CcdListAssembler is no longer
necessary.  Changed API for IsrTask.doCcdAssembly so it can cover both
options; this breaks use of this method with the 'methodList'.