interpolation: allow provision of fallback value
The interpolateFromMask and saturationCorrection functions did not allow
the provision of a fallback value for when the interpolation fails (and
did not provide one themselves).  The resulted in some pixels being set
to ~3.4e34 when the interpolation failed.

Added a fallbackValue keyword argument for these functions, and changed
them to use the interpolateDefectList function, which provides a
suitable fallbackValue when one is not provided.

Though this is an API change, it should be innocuous due to providing a
suitable default value.