overscan: do a single round of clipping when collapsing (#3065)
RHL pointed out that the overscan can include CR hits and signal
leaked from the main image.  It's necessary to remove these.
Do this by a single round of clipping, using the median and the
IQR.
Added a parameter to control the rejection level; formally changes
the API, but it's defaulted so shouldn't matter.