IsrTask: Add function to set valid Polygon
This is currently used only in the obs_subaru isr task.  It was put here
to make pushing to the LSST side easier.

Conflicts:
	python/lsst/ip/isr/isrTask.py

This is essentially a cherry-pick of HSC-976's:
https://github.com/HyperSuprime-Cam/ip_isr/commit/6739c63b91ee6bc89b74ff524d0b2a5870467cb7
but with several adaptations to LSST's API's.  Of note, the getCorners()
here returns the pixel coords at the actual pixel corner (i.e.
ll = (-0.5, -0.5) rather than (0, 0).  Thus the behaviour here is
slightly different than that on HSC, but this is actually the more
appropriate implementation.