Documentation update
Updated input documentation of overscan correction to reflect the various
options of types which may be passed into the function, as in practice an
image with a mask plane should be used, and not a simple an image. Also
updated code comments as to why a bit of code exists, i.e. it exists to
handle the case of extrapolation of saturated values in the overscan.