fringe: allow pedestal removal for multiple frames
In the run method, the input fringes can be a list of
fringe exposures.  removePedestal expects a single
fringe frame, so it could not work if a list of fringes
is given with config.pedestal on. This wasn't tested.

Edit the run method and rearrange using the placeholder
implementation for multi-mode fringe subtraction.
Add a test in testPedestal to test with multiple frames.