Add brighter fatter correction from obs_subaru as calibration option.
This implements the brighter fatter correction method from HSC.
We generated a kernel from flat field images that
can be applied iteratively to an image to correct for intensity
dependent PSFs.  We still need to investigate when to stop iterating,
but the current settings seem to work for HSC.

This change does not include the code to generate the kernel or test the algorithm.