Grow original mask when interpolating.
If we grow the region we interpolate when interpolating
masked pixels, we also need to grow the original mask so
we know why those pixels were interpolated; just setting
INTRP shouldn't be allowed.

This change refactors some code to make that possible,
including removing grow arguments from functions that could
not set the mask appropriately.