overscanCorrection: subtract fit from overscan
This allows downstream processing to inspect the fit results.

Because the overscan is now subtracted explicitly, this requires that the
overscan passed to this function not be contained within the image being
corrected (or the correction is applied twice). Unfortunately, it's not
clear how to detect this case; however, the primary user is IsrTask, which
pulls out the components explicitly and so shouldn't mix them.

Since this involved changes to the docstring, reformatted the docstring
for this function into numpydoc.