Convert isrTask to valid pipelineTask.
Additional configuration options now exist to define the required
input/output dataset types, and appropriate handlers correctly remove
these datasets if the isr configuration will not require (for inputs)
or generate (for outputs) those products.  As the camera configuration
is not currently fully populated, the obs_subaru/hsc configuration is
read to set the correct parameters.