Fix potential KeyError in gen3 datasets.
The inputTypeDict and outputTypeDict are constructed with all possible
input and output datasets, and are culled based on the configuration
parameters.  If the entry does not exist in the dict (possibly due to
other configurations), this can cause a KeyError.  Resolve this by
using dict.pop("key", None) instead of del dict["key"].