Allow use of new and old brighter-fatter kernels.
For gen2 ISR, move handling of new style (cp_pipe) and old style
(numpy.ndarray) brighter-fatter kernels entirely into
IsrTask.readIsrData() method.  Based on the configuration, this should
attempt to find the appropriate style, and optionally pull the kernel
for the detector to process.  Internally, the kernel is still passed
as the single numpy.ndarray needed for the processing.