Only mask high pixels during crosstalk correction.
With the change in background inclusion for crosstalk correction, the
manner in which pixels have the crosstalk bit set also needs to
change.  The masking threshold was previously set to an absolute pixel
level.  However, if the background level changes significantly, the
same field may result in only star peaks being masked (likely what is
expected) or the entire frame being masked (likely not).  Change the
threshold based on the full ccd background level, such that the
minPixelToMask is now relative to that background level.