Correct handling of crosstalk error/count/validity.
These elements were being silently dropped even when the input
calibration populated the entries.  To ensure this isn't the case,
initialize these fields as matching the shape of the coefficient table
when they do not exist.

For simplicity, the __init__ method now allows the calibration to be
initialized to the correct size if the nAmp parameter is passed.