Stop adjusting the raw WCS in assembleCcdTask
As of DM-20154, the WCS attahced to the raw exposure on read (and accessed
by rawExposure.getWcs()) is the best-guess raw WCS from the boresight +
rotation + cameraGeom (distortion).  As such, the WCS attached to the
inExposure here is already the "adjusted" WCS and the current call to
cameraGeomUtils.prepareWcsData() is actually performing an erroneous
adjustment.  This shifting has an effect downstream in the astrometry task
as the center sky coordinate used to define the reference object loading
sky circle is not actually our best guess.  For the most part, at least in
our HSC RC2 processing, the matching and fitting can handle this and a
decent WCS fit is found in single frame processing.  However, when probing
the full PDR2 dataset, many cases were revealed where this resulted in a
terrible WCS fit (with scatter on the order of 5 arcsec) and many others
where the ccd failed to process (often because, with the bad WCS fit, no
matches could be found for the photometry).