Add NO_DATA to masks list fed to brighter-fatter
The vignetted region of CCDs is marked by the NO_DATA mask plane.  Without
including NO_DATA in the list of masks passed to the brighter-fatter (BF)
correction task, it was possible to end up with non-finite values in the
image as a result of trying to apply these corrections in the vignetted
region (so, specifically limited to outer edge CCDs).  As NaN masking is
done before BF, these non-finite values were not getting marked as such
nor interpolated over but, more to the point, they never should have been
there in the first place.  With this addition, non-finite values are no
longer added during the BF corrections and thus no longer seen in the
processed images.

Since the mask list to be interpolated when applying the BF correction may
be different from the list of masks to permanently interpolate over, a new
BF-specific mask list config is added here and includes by default the
"SAT", "BAD", "NO_DATA", and "UNMASKEDNAN" masks.