Remove UNMASKEDNAN from default maskListToInterpolate
This mask is meant to be included in this list when doNanInterpolation is
True.  This condition is met in the config validation (i.e. it is added to
the list if not alerady there if doNanInterpolation is True), thus leaving
it off the default list allows more easily to run with doNanInterpolation
set to False.  As is, it is very error prone as the user would also have
to reset the default maskListToInterpolate to not include this plane.