Fix filter look-up after change to filterLabel
The function in isrFunctions to obtain the physical filter is to accommodate
filterLabels that do not have an associated physicaLabel attribute.  A current
situation where this arises is in mocked up images in a number of unittests
(those in pipe_tasks, in particular).  If no attribute exsists, the value
returned by this function is "Unknown".