Fix FULLCOVARIANCE PTC noise units to electron if PTCVER < 1.9
Before version 1.9, the noise stored in the PTC was in units
of electron^2 only if ptcFitType == FULLCOVARIANCE. After
version 1.9, we standardized the PhotonTransferCurveDataset
noise units to electron to fix this bug. If a user tries to
use an earlier version of PTC with this fit type, we must be
sure to do the calculations properly. More information about
this noise issue can be found in DM-45976.