Always generate interfaceOffsetDict
In the case whereby the entire image is masked, the interfaceOffsetDict
is not currently being generated. This causes a problem for a subsequent
for loop which sets the metadata for each interface ID. The change made
on this commit forces the task to always generate an interfaceOffsetDict
complete with interfaceIds. To facilitate this, it was necessary to add
an all NaN check inside getInterfaceOffset.